+++ tests/sheet.spec.tsx
@@ -1,80 +1,67 @@
-import dayjs from 'dayjs';
+import faker from 'faker';
+import XLSX from 'xlsx';
 
 import { Sheet } from '../src';
-
-const MOCK_DATA = [
-  ['name', 'height', 'weight'],
-  ['Joe', 188, 87],
-  ['Marry', 158, 54],
-  ['Bob', 177, 78],
-];
+import { mockAoaData, getMockedFormatDataTestCases } from './mocks';
 
 describe('Sheet class', () => {
-  const sheet = new Sheet('new_file', 'xlsx', MOCK_DATA);
+  const fakeName = faker.name.title();
+  const fakeData = mockAoaData();
+  const sheet = new Sheet(fakeName, fakeData);
   const sheetPrototype = Object.getPrototypeOf(sheet);
 
   it('throws error when it gets invalid fileName', () => {
+    const invalidName = '';
     const errorMessage = 'Invalid file name provided';
     expect(() => {
-      new Sheet('', 'xlsx', MOCK_DATA);
+      new Sheet(invalidName, fakeData);
     }).toThrow(errorMessage);
   });
 
-  describe('preprocessData', () => {
-    const cases = [
-      {
-        name: 'string',
-        before: [['string', 'hello']],
-        after: [['string', 'hello']],
-      },
-      { name: 'number', before: [['number', 123]], after: [['number', '123']] },
-      {
-        name: 'boolean',
-        before: [['boolean', true]],
-        after: [['boolean', 'TRUE']],
-      },
-      {
-        name: 'object',
-        before: [['object', { firstName: 'John', lastName: 'Doe' }]],
-        after: [['object', '{"firstName":"John","lastName":"Doe"}']],
-      },
-      {
-        name: 'date',
-        before: [['date', new Date()]],
-        after: [['date', dayjs(new Date()).format('YYYY-MM-DD')]],
-      },
-      {
-        name: 'undefined',
-        before: [['undefined', undefined]],
-        after: [['undefined', '']],
-      },
-      {
-        name: 'null',
-        before: [['null', null]],
-        after: [['null', 'null']],
-      },
-    ];
+  describe('validateData', () => {
+    it('should be throw error when it gets invalid data', () => {
+      const invalidData = [[]];
+      const errorMessage = 'Invalid data provided';
+      expect(() => {
+        new Sheet(fakeName, invalidData);
+      }).toThrow(errorMessage);
+    });
 
+    it('should be throw error when all rows are not a same length.', () => {
+      const invalidData = [
+        [1, 2, 3],
+        [4, 5],
+        [6, 7, 8, 9, 10],
+      ];
+      const errorMessage = 'All rows should be a same length';
+      expect(() => {
+        new Sheet(fakeName, invalidData);
+      }).toThrow(errorMessage);
+    });
+  });
+
+  describe('formatData', () => {
+    const cases = getMockedFormatDataTestCases();
     test.each(cases)('$name', ({ before, after }) => {
-      const preprocessedData = sheetPrototype.preprocessData(before);
-      expect(preprocessedData).toEqual(after);
+      const formattedData = sheetPrototype.formatData(before);
+      expect(formattedData).toEqual(after);
     });
   });
 
   it('creates workbook', () => {
-    const sheetName = 'my_excel_file';
-    const sheet = new Sheet(sheetName, 'xlsx', MOCK_DATA);
-    const sheetNames = sheet.createWorkBook().SheetNames;
-    expect(sheetName).toEqual(sheetNames[0]);
+    const workBook = sheetPrototype.createWorkBook(sheet.data);
+    const firstSheetName = workBook.SheetNames[0];
+    expect(firstSheetName).toEqual('Sheet1');
   });
 
-  it('joins name and extension', () => {
-    const result = sheetPrototype.joinNameAndExtension('new_file', 'xlsx');
-    expect(result).toEqual('new_file.xlsx');
+  it('formats full name', () => {
+    const result = sheetPrototype.formatFullName(fakeName, 'xlsx');
+    expect(result).toEqual(`${fakeName}.xlsx`);
   });
 
-  it('set options', () => {
-    sheetPrototype.setOptions({ dateFormat: 'YY/MM/DD' });
-    expect(sheetPrototype.options).toEqual({ dateFormat: 'YY/MM/DD' });
+  it('calls writeFile when download sheet', () => {
+    const spy = jest.spyOn(XLSX, 'writeFile').mockImplementation(jest.fn());
+    sheet.download();
+    expect(spy).toHaveBeenCalled();
   });
 });